Switch to work with booleans, convert to numbers before syncing
