+++ src/utils/TagUtils.ts
@@ -1,4 +1,5 @@
 import Log from '../libraries/Log';
+import { TagsObject } from '../models/Tags';
 
 export default class TagUtils {
     static async tagHelperWithRetries(callback:Function, ms: number, maxTries: number): Promise<Object> {
@@ -15,5 +16,12 @@ export default class TagUtils {
                 }
             });
         });
+    }
+
+    static convertTagBooleanValuesToNumbers(tags: TagsObject): TagsObject {
+        Object.keys(tags).forEach(key => {
+            tags[key] = Number(tags[key]);
+        });
+        return tags;
     }
 }
